Hound fixes #3 and localizations
