+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/@types/recorder.ts
@@ -34,15 +34,6 @@ export type UseRecorder = {
   saveRecording: () => void;
 };
 
-export type RecorderControlsProps = {
-  recorderState: Recorder;
-  handlers: {
-    startRecording: () => void;
-    cancelRecording: () => void;
-    saveRecording: () => void;
-  };
-};
-
 export type Audio = {
   key: string;
   audio: string;
