+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/actions/main-function/evaluation/evaluationActions.ts
@@ -1756,6 +1756,9 @@ const saveAssignmentEvaluationGradeToServer: SaveEvaluationAssignmentGradeEvalua
     };
   };
 
+/**
+ * saveAssignmentEvaluationSupplementationToServer
+ */
 const saveAssignmentEvaluationSupplementationToServer: SaveEvaluationAssignmentSupplementation =
   function saveAssignmentEvaluationSupplementationToServer({
     workspaceEntityId,
