+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/general/voice-recorder/recordings-list.tsx
@@ -1,11 +1,17 @@
 import * as React from "react";
 import { RecordValue } from "../../../@types/recorder";
-import Link from "~/components/general/link";
+import { StateType } from "../../../reducers/index";
+import { connect, Dispatch } from "react-redux";
+import { AnyActionType } from "../../../actions/index";
+import { bindActionCreators } from "redux";
+import Record from "./record";
+import { i18nType } from "../../../reducers/base/i18n";
 
 /**
  * RecordingsListProps
  */
 export interface RecordingsListProps {
+  i18n: i18nType;
   records: RecordValue[] | null;
   noDeleteFunctions?: boolean;
   deleteAudio?: (recordId: string) => void;
@@ -23,7 +29,7 @@ const defaultRecordListProps = {
  * @param props
  * @returns JSX.Element
  */
-export default function RecordingsList(props: RecordingsListProps) {
+function RecordingsList(props: RecordingsListProps) {
   props = { ...defaultRecordListProps, ...props };
 
   const { records, deleteAudio, noDeleteFunctions } = props;
@@ -31,26 +37,25 @@ export default function RecordingsList(props: RecordingsListProps) {
   return (
     <div className="voice__recorder-recordings__container">
       {records.length > 0 ? (
-        <>
-          <h1>Your recordings</h1>
-          <div className="voice__recorder-recordings__list">
-            {records.map((record) => {
-              return (
-                <Record
-                  controls
-                  record={record}
-                  src={record.url}
-                  key={record.id}
-                  noDeleteFunctions={noDeleteFunctions}
-                  onClickDelete={deleteAudio}
-                />
-              );
-            })}
-          </div>
-        </>
+        <div className="voice__recorder-recordings__list">
+          {records.map((record) => {
+            return (
+              <Record
+                controls
+                record={record}
+                src={record.url}
+                key={record.id}
+                noDeleteFunctions={noDeleteFunctions}
+                onClickDelete={deleteAudio}
+              />
+            );
+          })}
+        </div>
       ) : (
         <div className="voice__recorder-recordings__list voice__recorder-recordings__list--empty">
-          <span>You don't have records</span>
+          <span>
+            {props.i18n.text.get("plugin.workspace.audioField.noFiles")}
+          </span>
         </div>
       )}
     </div>
@@ -58,55 +63,21 @@ export default function RecordingsList(props: RecordingsListProps) {
 }
 
 /**
- * RecordProps
+ * mapStateToProps
+ * @param state
  */
-interface RecordProps
-  extends React.DetailedHTMLProps<
-    React.AudioHTMLAttributes<HTMLAudioElement>,
-    HTMLAudioElement
-  > {
-  record: RecordValue;
-  noDeleteFunctions: boolean;
-  onClickDelete?: (recordId: string) => void;
+function mapStateToProps(state: StateType) {
+  return {
+    i18n: state.i18n,
+  };
 }
 
 /**
- * defaultRecordtProps
+ * mapDispatchToProps
+ * @param dispatch
  */
-const defaultRecordtProps = {
-  noDeleteFunctions: false,
-};
-
-/**
- * Record
- * @param props
- * @returns JSX.Element
- */
-function Record(props: RecordProps) {
-  props = { ...defaultRecordtProps, ...props };
-  const audioRef = React.useRef(null);
-  const { record, onClickDelete, noDeleteFunctions, ...rest } = props;
-
-  return (
-    <div className="voice__recorder-recordings__list-item" key={rest.key}>
-      <audio ref={audioRef} {...rest} />
-      <div className="delete-button-container">
-        <Link
-          className="voice__recorder-audiofield-download-file-button icon-download"
-          title="lataa"
-          href={record.url}
-          openInNewTab={record.name}
-        />
-      </div>
-      {!noDeleteFunctions ? (
-        <div className="delete-button-container">
-          <Link
-            className="voice__recorder-audiofield-remove-file-button icon-trash"
-            title="Poista"
-            onClick={() => onClickDelete && onClickDelete(record.id)}
-          />
-        </div>
-      ) : null}
-    </div>
-  );
+function mapDispatchToProps(dispatch: Dispatch<AnyActionType>) {
+  return bindActionCreators({}, dispatch);
 }
+
+export default connect(mapStateToProps, mapDispatchToProps)(RecordingsList);
