+++ muikku-core-plugins/src/main/resources/fi/otavanopisto/muikku/plugins/evaluation/EvaluationJsPluginMessages_en.properties
@@ -39,6 +39,9 @@ plugin.evaluation.filter.viewSelection = Select View
 plugin.evaluation.evaluationModal.noJournals = No diary entries
 plugin.evaluation.evaluationModal.noAssignmentsTitle = No assignments
 
+plugin.evaluation.evaluationModal.audioAssessments = Audio assessments
+plugin.evaluation.evaluationModal.noAudioAssessments = No audio assessments
+
 plugin.evaluation.studentGrid.assessmentRequested.label = Assessment Requested
 plugin.evaluation.studentGrid.evaluated.label = Evaluated
 plugin.evaluation.studentGrid.joined.label = Joined
