+++ muikku-core-plugins/src/main/resources/fi/otavanopisto/muikku/plugins/evaluation/EvaluationJsPluginMessages_fi.properties
@@ -38,6 +38,8 @@ plugin.evaluation.filter.atWorkspace.filters = Suodattimet
 plugin.evaluation.filter.viewSelection = Rajaukset
 plugin.evaluation.evaluationModal.noJournals = Ei päiväkirjamerkintöjä
 plugin.evaluation.evaluationModal.noAssignmentsTitle = Ei tehtäviä
+plugin.evaluation.evaluationModal.audioAssessments = Äänipalauteet
+plugin.evaluation.evaluationModal.noAudioAssessments = Ei äänipalautteita
 
 plugin.evaluation.studentGrid.assessmentRequested.label = Arviointipyyntö
 plugin.evaluation.studentGrid.evaluated.label = Arvioitu
